requirements.txt
Pillow==9.2.0
rembg==2.0.30
tkinter==0.1.0